roadmap-generator/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.1
openai==2.16.0
python-dotenv==1.2.1
pydantic==2.12.5
